manual_control/requirements.txt
pygame 
opencv-python
pyserial